move C codegen into
